Feat: Add game logic
This feature adds game logic to process the current status of the game to ensure correct moves and win/loss state. This also includes an update to the game model where I replaced the game board arrays with an array that will store the individual moves in a format somewhat similar to standard chess notation (["Xa5", "Oe3", "Xc4"]). This allows us to 'replay' a game and I can possibly train some AI on the game data to make single player games more interesting that way.